优化代码，避免编译器告警

+++ include/mc/im/radix_tree/node.h
@@ -330,7 +330,7 @@ public:
     leaf_type    m_leaf;
     key_type     m_prefix;
     edges_type   m_edges;
-    size_t       m_version = 0;
+    int          m_version = 0;
     ref_ptr_type m_next;
     ref_ptr_type m_prev;
     alloc_type   m_alloc;
